requirements.txt
asgiref==3.8.1
Django==5.0.4
djangorestframework==3.15.1
pillow==10.3.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
sqlparse==0.5.0
django-filter==24.2
djangorestframework-simplejwt==5.3.1
pyjwt==2.8.0
coverage==7.5.1